User Login API

+++ backend/controllers/goalController.js
@@ -1,26 +1,28 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc   Get Goals
 // @route  GET /api/goals
 // @access Private 
 const getGoals = asyncHandler(async (req, res)=>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
 // @desc   Create/Set Goal
 // @route  POST /api/goals
 // @access Private 
-const setGoal = asyncHandler(async (req, res)=>{
+const setGoal = asyncHandler(async (req, res)=>{   
     if(!req.body) {
         res.status(400)
         throw new Error('Please add a text field');
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -34,6 +36,19 @@ const putGoal = asyncHandler(async (req, res)=>{
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //Check for user 
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make sure that logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -50,6 +65,19 @@ const deleteGoal = asyncHandler(async (req, res)=>{
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    
+    const user = await User.findById(req.user.id)
+
+    //Check for user 
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make sure that logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id, req.body, {new: false})
